6-size.c

+++ 0x00-hello_world/6-size.c
@@ -1,15 +1,15 @@
 #include<stdio.h>
 /**
- * main - Start
+ * main -  size of various types on the computer
  *
  * Return: Always 0 (Success)
  */
 int main(void)
 {
-	printf("Size of char: %lu byte(s)\n", sizeof(char));
-	printf("Size of int: %lu byte(s)\n", sizeof(int));
-	printf("Size of long int: %lu byte(s)\n", sizeof(long int));
-	printf("Size of long long: %lu byte(s)\n", sizeof(long long int));
-	printf("Size of float: %lu byte(s)\n", sizeof(float));
+	printf("Size of char: %d byte(s)\n", (int)sizeof(char));
+	printf("Size of int: %d byte(s)\n", (int)sizeof(int));
+	printf("Size of long int: %d byte(s)\n", (int)sizeof(long int));
+	printf("Size of long long: %d byte(s)\n", (int)sizeof(long long int));
+	printf("Size of float: %d byte(s)\n", (int)sizeof(float));
 	return (0);
 }
